Database_Covid/requirements.txt
numpy
streamlit
pandas
plotly
matplotlib
joblib
scikit-learn
streamlit-option-menu
CSV
Txt
option_menu
Pillow